// src/app/core/api/generated/model/storeFixedCostResponse.ts
/**
 * Precify API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { Periodicity } from './periodicity';


export interface StoreFixedCostResponse { 
    id?: string;
    storeId?: string;
    name?: string | null;
    category?: string | null;
    amount?: number;
    periodicity?: Periodicity;
    isActive?: boolean;
    createdAt?: string;
    updatedAt?: string;
}
export namespace StoreFixedCostResponse {
}