// src/app/core/api/generated/model/storeVariableCostResponse.ts
/**
 * Precify API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { VariableValueType } from './variableValueType';
import { VariableIncidenceType } from './variableIncidenceType';


export interface StoreVariableCostResponse { 
    id?: string;
    storeId?: string;
    salesChannelId?: string | null;
    name?: string | null;
    category?: string | null;
    valueType?: VariableValueType;
    amount?: number;
    incidenceType?: VariableIncidenceType;
    calculationOrder?: number | null;
    isActive?: boolean;
    createdAt?: string;
    updatedAt?: string;
}
export namespace StoreVariableCostResponse {
}