// src/app/core/api/generated/model/variableValueType.ts
/**
 * Precify API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export const VariableValueType = {
    Percentage: 'percentage',
    FixedValue: 'fixed_value'
} as const;
export type VariableValueType = typeof VariableValueType[keyof typeof VariableValueType];